Relocate repo-stats.ts into skill bundle (#238)
Move bin/architecture-overview/repo-stats.ts into skills/architecture-overview/scripts/
to comply with Anthropic skill anatomy. Executable code now lives inside the bundle
so the skill packages cleanly as a .skill file — scripts/package_skill.py no longer
ships a broken bundle with a dangling Step 3 invocation.

Updates:
- git mv bin/architecture-overview/repo-stats.ts → skills/architecture-overview/scripts/repo-stats.ts
- tests/architecture-overview.test.ts import path
- skills/architecture-overview/SKILL.md narrative (line 12) and Step 3 invocation (line 57)
- Removed empty bin/architecture-overview/ directory

No behavior change. Historical specs/prompts under docs/superpowers/ retain the old
path string by design (dated artifacts).
